refactor : chatting 페이지를 chatrooms로 변경 #59
